[AscendNPU-IR] fix build warnings

Created-by: jy834507063
Commit-by: jy834507063
Merged-by: tanshengshun
Description: ## 描述 Description
请提供此 Pull Request 的高级别说明。
Please describe what this PR is about.

## 类型 Category
- [ ] Bug修复 Bugfix
- [ ] 新功能 New feature
- [ ] 代码风格更新（格式化，局部变量）Style changes
- [ ] 重构（即不是新增功能，也不是修改bug的代码变动）Refactor
- [ ] 构建过程或辅助工具的变动 Build-related
- [ ] 文档内容更新 Docs

## Checklist
- [ ] 我的代码遵循这个项目的代码风格 My code follows the project's code style
- [ ] 我已经自己测试过我的代码 I have self-tested my code
- [ ] 我已经更新了相应的文档 I have updated the relevant documentation
- [ ] 我已经根据需要更新了对应的变更日志 I have updated the changelog as needed
- [ ] 我已经在标题中正确使用了类型标签（例如：`feat:`, `fix:`）I have correctly used type labels in the title (e.g., feat:, fix:)



See merge request: Ascend/AscendNPU-IR!965

+++ bishengir/lib/Dialect/HIVM/Transforms/GraphSyncSolver/SyncSolverTester.cpp
@@ -367,8 +367,8 @@ llvm::LogicalResult SyncTester::runSimulation(int runId, bool debugPrint) {
                corePipe);
         auto &triggeredOps = triggeredSetFlagOps[getTriggeredGroup(waitFlagOp)];
         assert(!waitFlagOp->eventIds.empty());
-        auto eventId =
-            waitFlagOp->eventIds[loopIdx % waitFlagOp->eventIds.size()];
+        auto eventId = waitFlagOp->eventIds[loopIdx % static_cast<int>(
+            waitFlagOp->eventIds.size())];
         auto it = triggeredOps.find(eventId);
         if (it != triggeredOps.end()) {
           assert((*it) == eventId);
@@ -559,7 +559,8 @@ llvm::LogicalResult SyncTester::runSimulation(int runId, bool debugPrint) {
     }
     assert(!setFlagOp->eventIds.empty());
     triggeredSetFlagOps[getTriggeredGroup(setFlagOp)].insert(
-        setFlagOp->eventIds[loopIndex % setFlagOp->eventIds.size()]);
+        setFlagOp->eventIds[loopIndex %
+                            static_cast<int>(setFlagOp->eventIds.size())]);
     refreshPipeQue(corePipeDst);
   };
